Layer Documentation Refactored
