+++ tensorlayer/decorators/__init__.py
@@ -9,7 +9,7 @@ layers that allow user to apply ``tf.nn.lrn`` on ``network.outputs``.
 More functions can be found in `TensorFlow API <https://www.tensorflow.org/versions/master/api_docs/index.html>`__.
 """
 
-from .deprecation import deprecated_alias
+from .deprecated_alias import deprecated_alias
 from .method_decorator import private_method
 from .method_decorator import protected_method
