+++ tensorlayer/decorators/deprecated_alias.py
