+++ tensorlayer/layers/image_resampling.py
