+++ tensorlayer/layers/importer.py
@@ -12,76 +12,10 @@ from tensorlayer import tl_logging as logging
 from tensorlayer.decorators import deprecated_alias
 
 __all__ = [
-    'LambdaLayer',
     'SlimNetsLayer',
     'KerasLayer',
     'EstimatorLayer',
 ]
-
-
-class LambdaLayer(Layer):
-    """A layer that takes a user-defined function using TensorFlow Lambda, for multiple inputs see :class:`ElementwiseLambdaLayer`.
-
-    Parameters
-    ----------
-    prev_layer : :class:`Layer`
-        Previous layer.
-    fn : function
-        The function that applies to the outputs of previous layer.
-    fn_args : dictionary or None
-        The arguments for the function (option).
-    name : str
-        A unique layer name.
-
-    Examples
-    ---------
-    Non-parametric case
-
-    >>> import tensorflow as tf
-    >>> import tensorlayer as tl
-    >>> x = tf.placeholder(tf.float32, shape=[None, 1], name='x')
-    >>> net = tl.layers.InputLayer(x, name='input')
-    >>> net = tl.layers.LambdaLayer(net, lambda x: 2*x, name='lambda')
-
-    Parametric case, merge other wrappers into TensorLayer
-
-    >>> from keras.layers import *
-    >>> from tensorlayer.layers import *
-    >>> def keras_block(x):
-    >>>     x = Dropout(0.8)(x)
-    >>>     x = Dense(800, activation='relu')(x)
-    >>>     x = Dropout(0.5)(x)
-    >>>     x = Dense(800, activation='relu')(x)
-    >>>     x = Dropout(0.5)(x)
-    >>>     logits = Dense(10, activation='linear')(x)
-    >>>     return logits
-    >>> net = InputLayer(x, name='input')
-    >>> net = LambdaLayer(net, fn=keras_block, name='keras')
-
-    """
-
-    @deprecated_alias(layer='prev_layer', end_support_version=1.9)  # TODO remove this line for the 1.9 release
-    def __init__(
-            self,
-            prev_layer,
-            fn,
-            fn_args=None,
-            name='lambda_layer',
-    ):
-
-        super(LambdaLayer, self).__init__(prev_layer=prev_layer, fn_args=fn_args, name=name)
-
-        logging.info("LambdaLayer  %s" % self.name)
-
-        if fn is None:
-            raise AssertionError("The `fn` argument cannot be None")
-
-        with tf.variable_scope(name) as vs:
-            self.outputs = fn(self.inputs, **self.fn_args)
-            variables = tf.get_collection(TF_GRAPHKEYS_VARIABLES, scope=vs.name)
-
-        self._add_layers(self.outputs)
-        self._add_params(variables)
 
 
 class SlimNetsLayer(Layer):
@@ -147,7 +81,6 @@ class SlimNetsLayer(Layer):
         self._add_params(slim_variables)
 
 
-@deprecated("2018-06-30", "This layer will be deprecated soon as :class:`LambdaLayer` can do the same thing.")
 class KerasLayer(Layer):
     """A layer to import Keras layers into TensorLayer.
 
@@ -167,6 +100,7 @@ class KerasLayer(Layer):
     """
 
     @deprecated_alias(layer='prev_layer', end_support_version=1.9)  # TODO remove this line for the 1.9 release
+    @deprecated("2018-06-30", "This layer will be deprecated soon as :class:`LambdaLayer` can do the same thing.")
     def __init__(
             self,
             prev_layer,
@@ -189,7 +123,6 @@ class KerasLayer(Layer):
         self._add_params(variables)
 
 
-@deprecated("2018-06-30", "This layer will be deprecated soon as :class:`LambdaLayer` can do the same thing.")
 class EstimatorLayer(Layer):
     """A layer that accepts a user-defined model.
 
@@ -211,6 +144,7 @@ class EstimatorLayer(Layer):
     @deprecated_alias(
         layer='prev_layer', args='layer_args', end_support_version=1.9
     )  # TODO remove this line for the 1.9 release
+    @deprecated("2018-06-30", "This layer will be deprecated soon as :class:`LambdaLayer` can do the same thing.")
     def __init__(
             self,
             prev_layer,
